serve webdav: address review feedback on Overwrite default
Move the Overwrite regression tests from overwrite_test.go into
webdav_test.go, reusing that file's server setup machinery, and drop the
separate file per reviewer request.

Trim the narrative comment in webdav.go to a concise pointer at the
upstream x/net/webdav bug (golang/go#66059).
