+++ cmd/serve/webdav/webdav.go
@@ -413,14 +413,9 @@ func (w *WebDAV) ServeHTTP(rw http.ResponseWriter, r *http.Request) {
 		w.serveDir(rw, r, remote)
 		return
 	}
-	// RFC 4918 section 10.6: when the Overwrite header is omitted from a
-	// COPY or MOVE request the resource MUST treat the request as if
-	// Overwrite: T had been sent. The upstream golang.org/x/net/webdav
-	// library mis-handles this for MOVE by checking == "T" instead of
-	// != "F", so an absent header is treated as Overwrite: F and the
-	// request fails with 412 Precondition Failed. Normalise the header
-	// here when the client did not send one, so the default matches the
-	// RFC for both verbs. See https://github.com/rclone/rclone/issues/9496.
+	// Work around bug in x/net/webdav which handles Overwrite incorrectly
+	// See: https://github.com/golang/go/issues/66059
+	// Remove when the above bug is fixed.
 	if (r.Method == "COPY" || r.Method == "MOVE") && r.Header.Get("Overwrite") == "" {
 		r.Header.Set("Overwrite", "T")
 	}
